refactor: clean up code and gitignore
