+++ sdks/rust/.gitignore
@@ -6,7 +6,3 @@
 /client/test
 /client/.gitignore
 /client/LICENSE
-/client/docs
-/client/src
-/target/
-/client/
